requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==5.4.0
telegram==0.0.1
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
